feat: add a network settings hook in the feature toggle saga
The update is called periodically by the monitor, that calls
unleash and overwrite the feature toggles.

To avoide an overwrite over the feature toggles when access
the wallet service is impossible due to the lack of the
wallet service in the custom network, we need this hook
that verifies if the wallet service is present, if not
it disables the wallet service and push notification features.